separate validation logic from licenceService
